restructured entity folder
