+++ src/main/java/jMonkeyEngine/Entities/Cars/CarComponents/Tire/Tire01.java
@@ -1,4 +1,4 @@
-package jMonkeyEngine.Entities;
+package jMonkeyEngine.Entities.Cars.CarComponents.Tire;
 
 import java.util.logging.Logger;
