+++ src/main/java/jMonkeyEngine/Entities/Cars/VehicleModels/Nismo.java
@@ -1,10 +1,15 @@
-package jMonkeyEngine.Entities;
+package jMonkeyEngine.Entities.Cars.VehicleModels;
 
 import com.jme3.asset.AssetManager;
 import com.jme3.bullet.PhysicsSpace;
-import com.jme3.bullet.control.VehicleControl;
 import com.jme3.math.Vector3f;
-import java.util.logging.Level;
+import jMonkeyEngine.Entities.Cars.CarComponents.*;
+import jMonkeyEngine.Entities.Cars.CarComponents.Steering;
+import jMonkeyEngine.Entities.Cars.CarComponents.Suspension;
+import jMonkeyEngine.Entities.Cars.CarComponents.Tire.Tire01;
+import jMonkeyEngine.Entities.Cars.CarComponents.Wheel.DarkAlloyWheel;
+import jMonkeyEngine.Entities.Cars.CarComponents.Wheel.Wheel;
+import jMonkeyEngine.Entities.Cars.CarComponents.Wheel.WheelModel;
 import java.util.logging.Logger;
 
 /**
