+++ src/main/java/jMonkeyEngine/Entities/Cars/VehicleModels/Vehicle.java
@@ -1,33 +1,22 @@
-package jMonkeyEngine.Entities;
+package jMonkeyEngine.Entities.Cars.VehicleModels;
 
-import com.jme3.anim.AnimComposer;
-import com.jme3.app.state.AppStateManager;
 import com.jme3.asset.AssetManager;
 import com.jme3.asset.AssetNotFoundException;
-import com.jme3.bounding.BoundingBox;
 import com.jme3.bullet.PhysicsSpace;
 import com.jme3.bullet.PhysicsTickListener;
-import com.jme3.bullet.animation.RagUtils;
-import com.jme3.bullet.collision.PhysicsCollisionObject;
-import com.jme3.bullet.collision.PhysicsRayTestResult;
 import com.jme3.bullet.collision.shapes.CollisionShape;
 import com.jme3.bullet.control.VehicleControl;
-import com.jme3.bullet.joints.PhysicsJoint;
-import com.jme3.bullet.objects.PhysicsRigidBody;
-import com.jme3.bullet.objects.PhysicsVehicle;
 import com.jme3.bullet.objects.VehicleWheel;
-import com.jme3.bullet.objects.infos.RigidBodyMotionState;
 import com.jme3.bullet.util.CollisionShapeFactory;
-import com.jme3.math.FastMath;
 import com.jme3.math.Matrix3f;
-import com.jme3.math.Quaternion;
 import com.jme3.math.Transform;
 import com.jme3.math.Vector3f;
-import com.jme3.renderer.queue.RenderQueue;
 import com.jme3.scene.Node;
 import com.jme3.scene.Spatial;
-import java.lang.reflect.Constructor;
-import java.lang.reflect.InvocationTargetException;
+import jMonkeyEngine.Entities.Cars.CarComponents.Steering;
+import jMonkeyEngine.Entities.Cars.CarComponents.Suspension;
+import jMonkeyEngine.Entities.Cars.CarComponents.Wheel.Wheel;
+import jMonkeyEngine.Entities.Cars.CarComponents.Wheel.WheelModel;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.logging.Logger;
